Switch to use CoreLib Path class.
Leaving the code for UAP for now- we may end up needing to keep it.
(Notably for the Crypto)

Use the core environment variable code.
