+++ src/System.Runtime.Extensions/src/System/Environment.Win32.cs
@@ -47,165 +47,6 @@ namespace System
             return StringBuilderCache.GetStringAndRelease(result);
         }
 
-        private static string GetEnvironmentVariableCore(string variable)
-        {
-            StringBuilder sb = StringBuilderCache.Acquire(128); // a somewhat reasonable default size
-            int requiredSize = Interop.Kernel32.GetEnvironmentVariableW(variable, sb, sb.Capacity);
-            if (requiredSize == 0 && Marshal.GetLastWin32Error() == Interop.Errors.ERROR_ENVVAR_NOT_FOUND)
-            {
-                StringBuilderCache.Release(sb);
-                return null;
-            }
-
-            while (requiredSize > sb.Capacity)
-            {
-                sb.Capacity = requiredSize;
-                sb.Length = 0;
-                requiredSize = Interop.Kernel32.GetEnvironmentVariableW(variable, sb, sb.Capacity);
-            }
-
-            return StringBuilderCache.GetStringAndRelease(sb);
-        }
-
-        private static string GetEnvironmentVariableCore(string variable, EnvironmentVariableTarget target)
-        {
-            if (target == EnvironmentVariableTarget.Process)
-            {
-                return GetEnvironmentVariableCore(variable);
-            }
-            else
-            {
-                RegistryKey baseKey;
-                string keyName;
-
-                if (target == EnvironmentVariableTarget.Machine)
-                {
-                    baseKey = Registry.LocalMachine;
-                    keyName = @"System\CurrentControlSet\Control\Session Manager\Environment";
-                }
-                else
-                {
-                    Debug.Assert(target == EnvironmentVariableTarget.User);
-                    baseKey = Registry.CurrentUser;
-                    keyName = "Environment";
-                }
-
-                using (RegistryKey environmentKey = baseKey.OpenSubKey(keyName, writable: false))
-                {
-                    return environmentKey?.GetValue(variable) as string;
-                }
-            }
-        }
-
-        private static IDictionary GetEnvironmentVariablesCore()
-        {
-            // Format for GetEnvironmentStrings is:
-            // (=HiddenVar=value\0 | Variable=value\0)* \0
-            // See the description of Environment Blocks in MSDN's
-            // CreateProcess page (null-terminated array of null-terminated strings).
-            // Note the =HiddenVar's aren't always at the beginning.
-
-            // Copy strings out, parsing into pairs and inserting into the table.
-            // The first few environment variable entries start with an '='.
-            // The current working directory of every drive (except for those drives
-            // you haven't cd'ed into in your DOS window) are stored in the 
-            // environment block (as =C:=pwd) and the program's exit code is 
-            // as well (=ExitCode=00000000).
-
-            var results = new LowLevelDictionary<string, string>();
-            char[] block = GetEnvironmentCharArray();
-            for (int i = 0; i < block.Length; i++)
-            {
-                int startKey = i;
-
-                // Skip to key. On some old OS, the environment block can be corrupted. 
-                // Some will not have '=', so we need to check for '\0'. 
-                while (block[i] != '=' && block[i] != '\0') i++;
-                if (block[i] == '\0') continue;
-
-                // Skip over environment variables starting with '='
-                if (i - startKey == 0)
-                {
-                    while (block[i] != 0) i++;
-                    continue;
-                }
-
-                string key = new string(block, startKey, i - startKey);
-                i++;  // skip over '='
-
-                int startValue = i;
-                while (block[i] != 0) i++; // Read to end of this entry 
-                string value = new string(block, startValue, i - startValue); // skip over 0 handled by for loop's i++
-
-                results[key] = value;
-            }
-            return results;
-        }
-
-        private static IDictionary GetEnvironmentVariablesCore(EnvironmentVariableTarget target)
-        {
-            if (target == EnvironmentVariableTarget.Process)
-            {
-                return GetEnvironmentVariablesCore();
-            }
-            else
-            {
-                RegistryKey baseKey;
-                string keyName;
-                if (target == EnvironmentVariableTarget.Machine)
-                {
-                    baseKey = Registry.LocalMachine;
-                    keyName = @"System\CurrentControlSet\Control\Session Manager\Environment";
-                }
-                else
-                {
-                    Debug.Assert(target == EnvironmentVariableTarget.User);
-                    baseKey = Registry.CurrentUser;
-                    keyName = @"Environment";
-                }
-
-                using (RegistryKey environmentKey = baseKey.OpenSubKey(keyName, writable: false))
-                {
-                    var table = new LowLevelDictionary<string, string>();
-                    if (environmentKey != null)
-                    {
-                        foreach (string name in environmentKey.GetValueNames())
-                        {
-                            table.Add(name, environmentKey.GetValue(name, "").ToString());
-                        }
-                    }
-                    return table;
-                }
-            }
-        }
-
-        private static unsafe char[] GetEnvironmentCharArray()
-        {
-            // Format for GetEnvironmentStrings is:
-            // [=HiddenVar=value\0]* [Variable=value\0]* \0
-            // See the description of Environment Blocks in MSDN's
-            // CreateProcess page (null-terminated array of null-terminated strings).
-            char* pStrings = Interop.Kernel32.GetEnvironmentStringsW();
-            if (pStrings == null)
-            {
-                throw new OutOfMemoryException();
-            }
-            try
-            {
-                // Search for terminating \0\0 (two unicode \0's).
-                char* p = pStrings;
-                while (!(*p == '\0' && *(p + 1) == '\0')) p++;
-
-                var block = new char[(int)(p - pStrings + 1)];
-                Marshal.Copy((IntPtr)pStrings, block, 0, block.Length);
-                return block;
-            }
-            finally
-            {
-                Interop.Kernel32.FreeEnvironmentStringsW(pStrings); // ignore any cleanup error
-            }
-        }
-
         private static string GetFolderPathCore(SpecialFolder folder, SpecialFolderOption option)
         {
             // We're using SHGetKnownFolderPath instead of SHGetFolderPath as SHGetFolderPath is
@@ -478,82 +319,12 @@ namespace System
             return 0;
         });
 
-        private static void SetEnvironmentVariableCore(string variable, string value)
-        {
-            if (!Interop.Kernel32.SetEnvironmentVariableW(variable, value))
-            {
-                int errorCode = Marshal.GetLastWin32Error();
-                switch (errorCode)
-                {
-                    case Interop.Errors.ERROR_ENVVAR_NOT_FOUND: // Allow user to try to clear a environment variable
-                        return;
-                    case Interop.Errors.ERROR_FILENAME_EXCED_RANGE: // Fix inaccurate error code from Win32
-                        throw new ArgumentException(SR.Argument_LongEnvVarValue, nameof(value));
-                    default:
-                        throw new ArgumentException(Interop.Kernel32.GetMessage(errorCode));
-                }
-            }
-        }
-
-        private static void SetEnvironmentVariableCore(string variable, string value, EnvironmentVariableTarget target)
-        {
-            if (target == EnvironmentVariableTarget.Process)
-            {
-                SetEnvironmentVariableCore(variable, value);
-            }
-            else
-            {
-                RegistryKey baseKey;
-                string keyName;
-
-                if (target == EnvironmentVariableTarget.Machine)
-                {
-                    baseKey = Registry.LocalMachine;
-                    keyName = @"System\CurrentControlSet\Control\Session Manager\Environment";
-                }
-                else
-                {
-                    Debug.Assert(target == EnvironmentVariableTarget.User);
-
-                    // User-wide environment variables stored in the registry are limited to 255 chars for the environment variable name.
-                    const int MaxUserEnvVariableLength = 255;
-                    if (variable.Length >= MaxUserEnvVariableLength)
-                    {
-                        throw new ArgumentException(SR.Argument_LongEnvVarValue, nameof(variable));
-                    }
-
-                    baseKey = Registry.CurrentUser;
-                    keyName = "Environment";
-                }
-
-                using (RegistryKey environmentKey = baseKey.OpenSubKey(keyName, writable: true))
-                {
-                    if (environmentKey != null)
-                    {
-                        if (value == null)
-                        {
-                            environmentKey.DeleteValue(variable, throwOnMissingValue: false);
-                        }
-                        else
-                        {
-                            environmentKey.SetValue(variable, value);
-                        }
-                    }
-                }
-            }
-
-            //// Desktop sends a WM_SETTINGCHANGE message to all windows.  Not available on all platforms.
-            //Interop.Kernel32.SendMessageTimeout(
-            //    new IntPtr(Interop.Kernel32.HWND_BROADCAST), Interop.Kernel32.WM_SETTINGCHANGE,
-            //    IntPtr.Zero, "Environment", 0, 1000, IntPtr.Zero);
-        }
-
         public static string SystemDirectory
         {
             get
             {
-                StringBuilder sb = StringBuilderCache.Acquire(Path.MaxPath);
-                if (Interop.Kernel32.GetSystemDirectoryW(sb, Path.MaxPath) == 0)
+                StringBuilder sb = StringBuilderCache.Acquire(PathInternal.MaxShortPath);
+                if (Interop.Kernel32.GetSystemDirectoryW(sb, PathInternal.MaxShortPath) == 0)
                 {
                     StringBuilderCache.Release(sb);
                     throw Win32Marshal.GetExceptionForLastWin32Error();
